SessionUtils 파일 추가 - 세션 공통 처리

+++ src/main/java/spring/board/service/BoardService.java
@@ -13,6 +13,7 @@ import spring.board.dto.common.PageInfo;
 import spring.board.dto.file.FileRequest;
 import spring.board.dto.file.FileResponse;
 import spring.board.dto.user.UserSession;
+import spring.board.util.SessionUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
@@ -23,7 +24,6 @@ import java.util.*;
 @Service
 //@RequiredArgsConstructor
 public class BoardService {
-    private final UserService userService;
     private final FileService fileService;
     private final CommentService commentService;
     private final BoardMapper boardMapper;
@@ -32,8 +32,7 @@ public class BoardService {
     //    private final JdbcBoardDao boardDao;
 //    private final JdbcFileDao fileDao;
 
-    public BoardService(UserService userService, FileService fileService, CommentService commentService, JdbcBoardDao boardDao, JdbcFileDao fileDao, JdbcCommentDao commentDao, BoardMapper boardMapper, FileMapper fileMapper) {
-        this.userService = userService;
+    public BoardService(FileService fileService, CommentService commentService, BoardMapper boardMapper, FileMapper fileMapper) {
         this.fileService = fileService;
         this.commentService = commentService;
         this.fileMapper = fileMapper;
@@ -45,9 +44,10 @@ public class BoardService {
 
 
 
-    public String post(BoardRequest boardRequest, HttpServletRequest request) throws IOException {
-        HttpSession session = request.getSession();
-        UserSession userSession = (UserSession) session.getAttribute("USER");
+    public String post(BoardRequest boardRequest, HttpServletRequest request) throws Exception {
+//        HttpSession session = request.getSession();
+//        UserSession userSession = (UserSession) session.getAttribute("USER");
+        UserSession userSession = (UserSession) SessionUtils.getAttribute("USER");
 
         if (userSession != null) {
             Board board = new Board(null, boardRequest.getTitle(), boardRequest.getContent(), userSession.getIdx(), new Timestamp(new Date().getTime()), null);
@@ -106,15 +106,11 @@ public class BoardService {
         return boardInfoResponse;
     }
 
-    public List<BoardResponse> selectPostAndCommentByPostId(Integer bIdx) {
-        BoardResponse boardResponse = new BoardResponse();
-        return null;
-    }
-
     // 내 게시물 - 사용자 검색
-    public List<BoardResponse> selectPostsByUserId(int page, int size, int blockSize, HttpServletRequest request) {
-        HttpSession session = request.getSession();
-        UserSession userSession = (UserSession) session.getAttribute("USER");
+    public List<BoardResponse> selectPostsByUserId(int page, int size, int blockSize) throws Exception {
+//        HttpSession session = request.getSession();
+//        UserSession userSession = (UserSession) session.getAttribute("USER");
+        UserSession userSession = (UserSession) SessionUtils.getAttribute("USER");
 
         if (userSession != null) {
             Paging paging = new Paging(page, size, blockSize, boardMapper.getTotalCnt());
@@ -131,43 +127,57 @@ public class BoardService {
         }
     }
 
-    public String updatePost(Integer bIdx, BoardRequest boardRequest, HttpServletRequest request) {
-        UserSession userSession = userService.getLoginUserInfo(request);
-        Integer b_uidx = boardMapper.selectPostByPostId(bIdx).getUserIdx();
+    public String updatePost(Integer bIdx, BoardRequest boardRequest) throws Exception {
+//        UserSession userSession = userService.getLoginUserInfo(request);
+        UserSession userSession = (UserSession) SessionUtils.getAttribute("USER");
 
-        if (userSession != null ) {
-            if (b_uidx.equals(userSession.getIdx())) {
-                BoardUpdateRequest boardUpdateRequest = new BoardUpdateRequest(bIdx, boardRequest.getTitle(), boardRequest.getContent(), new Timestamp(new Date().getTime()));
-                boardMapper.updatePost(boardUpdateRequest);
-                return "게시물 수정 완료";
+        // mybatis에서 본인 idx와 게시물 idx가 같으면
+        BoardResponse board = boardMapper.selectPostByPostId(bIdx);
+
+        if (board != null) {
+            Integer b_uidx = board.getUserIdx();
+
+            if (userSession != null ) {
+                if (b_uidx.equals(userSession.getIdx())) {
+                    Date now = new Date();
+                    BoardUpdateRequest boardUpdateRequest = new BoardUpdateRequest(bIdx, boardRequest.getTitle(), boardRequest.getContent(), new Date());
+                    boardMapper.updatePost(boardUpdateRequest);
+                    return "게시물 수정 완료";
+                } else {
+                    System.out.println("본인 게시물만 삭제 가능");
+                }
             } else {
-                System.out.println("본인 게시물만 삭제 가능");
+                System.out.println("사용자 로그인 정보 없음");
             }
         } else {
-            System.out.println("사용자 로그인 정보 없음");
+            System.out.println("게시물이 존재하지 않음");
         }
         return "게시물 수정 실패";
     }
 
-    public String deletePost(Integer bIdx, HttpServletRequest request) {
+    public String deletePost(Integer bIdx) throws Exception {
 
-        UserSession userSession = userService.getLoginUserInfo(request);
-        Integer b_uidx = boardMapper.selectPostByPostId(bIdx).getUserIdx();
+        // MyBatis에서 본인의 idx와 idx가 같다면
+//        UserSession userSession = userService.getLoginUserInfo(request);
+        UserSession userSession = (UserSession) SessionUtils.getAttribute("USER");
+        BoardResponse board = boardMapper.selectPostByPostId(bIdx);
 
-        if (userSession != null ) {
-            if (b_uidx.equals(userSession.getIdx())) {
-                boardMapper.deletePost(bIdx);
-                return "게시물 삭제 완료";
+        if (board != null) {
+            Integer b_uidx = boardMapper.selectPostByPostId(bIdx).getUserIdx();
+
+            if (userSession != null ) {
+                if (b_uidx.equals(userSession.getIdx())) {
+                    boardMapper.deletePost(bIdx);
+                    return "게시물 삭제 완료";
+                } else {
+                    System.out.println("본인 게시물만 삭제 가능");
+                }
             } else {
-                System.out.println("본인 게시물만 삭제 가능");
+                System.out.println("사용자 로그인 정보 없음");
             }
         } else {
-            System.out.println("사용자 로그인 정보 없음");
+            System.out.println("게시물이 존재하지 않음");
         }
         return "게시물 삭제 실패";
-    }
-
-    public BoardResponse searchPosts(String q) {
-        return boardMapper.searchPosts(q);
     }
 }
